+++ native/src/seal/ckks.h
@@ -728,8 +728,8 @@ namespace seal
             destination.reserve(slots_);
             for (std::size_t i = 0; i < slots_; i++)
             {
-                destination.emplace_back(
-                    from_complex<T>(res[matrix_reps_index_map_[i]]));
+                destination.emplace_back(from_complex<T>(
+                    res[static_cast<std::size_t>(matrix_reps_index_map_[i])]));
             }
         }
 
@@ -757,6 +757,6 @@ namespace seal
 
         util::Pointer<std::complex<double>> inv_roots_;
 
-        util::Pointer<std::uint64_t> matrix_reps_index_map_;
+        util::Pointer<std::size_t> matrix_reps_index_map_;
     };
 }
